+++ CHANGELOG.md
@@ -6,6 +6,7 @@ MAGE adheres to [Semantic Versioning](http://semver.org/).
 ## Pending on [`develop`](https://github.com/ngageoint/mage-server/tree/develop)
 
 ##### Features
+* Users can view or edit coordinates as MGRS.
 
 ##### Bug Fixes